Allow collection after ^C.
Before this change, the script would only clean up background processes,
leaving unreadable trace files, which customers would send back to us.
This just wastes a lot of time - the traces are huge, but, essentially
corrupted - trace-cmd report can't read them and errors out.

With this change the journal event catched runs in the background again,
and we store it's PID. It writes to an event file in case an event
happens. So we will have a copy of the event that occurred in the dist
tarball, which is a net positive. If the user ^C's the script, we
SIGTERM the journal subtask, and manually write something into the event
file to signal that ^C happened, and let the script wrap up all tracing
as it should have.

This now correctly tarballs up traces, collects dmesg and event copies,
and compresses the whole folder for collection.

I've added an idle spinner in the main waiting thread, just so users are
aware that something is happening in the terminal in the background.

Signed-off-by: Auke Kok <auke.kok@versity.com>