Enable restoring truncated trace files.
With `trace-cmd restore` we can potentially restore trace output
files that were not properly collated because of an IPMI reboot, KILL or
other interrupt.

However, trace-cmd needs some data from the system itself where it is
run, or else it won't be able to decypher the trace events, and the
trace will remain unreadable.

So, we can capture this `partial.dat` right as we trace at all times,
effectively making the capture trace safe from reboots. As long as the
trace buffer data was spooled to disk, we can reassemble it properly.

Use `trace-cmd -i partial.dat trace.dat.cpu*` to do this reassembling.

Signed-off-by: Auke Kok <auke.kok@versity.com>